feat:make copy

+++ src/fs/create.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs  from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
@@ -7,14 +7,18 @@ const __dirname = dirname(__filename);
 
 export const create = async () => {
     const filePath= path.join(__dirname, 'files', 'fresh.txt');
-    if (fs.existsSync(filePath)) {
-        console.error('FS operation failed');
-    }
-    fs.writeFile(filePath, 'I am fresh and young', (error) => {
-        if (error) {
-            throw new Error();
+    fs.open(filePath, 'wx', (err, fd) => {
+        if (err && err.code === 'EEXIST') {
+            throw new Error('FS operation failed');
+            return;
         }
+        fs.writeFile(filePath, 'I am fresh and young', (error) => {
+            if (error) {
+                throw new Error();
+            }
+        });
     });
+
 };
 
 create();
